requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
aiofiles==23.2.1
python-dotenv==1.0.1